backend/requirements.txt
# バックエンド（Django/DRF）で使う Python 依存関係です。
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
drf-spectacular>=0.27
drf-spectacular-sidecar>=2024.0.0
djangorestframework-simplejwt>=5.3
